requirements.txt
django
gunicorn
dj-database-url
psycopg2-binary
whitenoise
django-filter
markdown
djangorestframework
drf-yasg[validation]
django-bootstrap4
django-auth-ldap

